Patch from Stefan Fuhrmann (polarfox@t-online.de): Make TSVN compatible with 32 and 64 bit WinAPI. Also provide some fixes for "paranoid" 64 bit compatibility warnings.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@1774 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseShell/ShellCache.h
@@ -263,7 +263,7 @@ private:
 				return;
 			excludeliststr = (stdstring)excludelist;
 			exvector.clear();
-			int pos = 0, pos_ant = 0;
+			size_t pos = 0, pos_ant = 0;
 			pos = excludeliststr.find(_T("\n"), pos_ant);
 			while (pos != stdstring::npos)
 			{
@@ -289,7 +289,7 @@ private:
 				return;
 			includeliststr = (stdstring)includelist;
 			invector.clear();
-			int pos = 0, pos_ant = 0;
+			size_t pos = 0, pos_ant = 0;
 			pos = includeliststr.find(_T("\n"), pos_ant);
 			while (pos != stdstring::npos)
 			{
